Convert hui-view to Lit (#2265)
* Convert hui-view to Lit

* Add super call to updated

* Update src/panels/lovelace/hui-view.ts


* Apply suggestions from code review


* Address comments"

* Fix things
